refactor(rs-drive-proof-verifier): move some code around
