+++ packages/rs-drive-proof-verifier/src/verify.rs
@@ -8,7 +8,7 @@ use tenderdash_abci::{
 
 use crate::Error;
 
-use crate::proof::QuorumInfoProvider;
+use crate::QuorumInfoProvider;
 
 /// Verify cryptographic proof generated by Tenderdash
 ///
